Responsiveness of Bluesky link in SocialNetwor
The `bluesky` button in the `sections/socialNetwork.html` partial wasn't
resizing correctly when the document `fontSize` was changed (i.e. in
`hugo.yaml`). This is because the underlying svg was being defined with
an absolute height and width in `px`, rather than a relative size in
`em`.

The root cause is that this link was patterned after the Twitter link,
which *also* doesn't behave correctly with changing root `fontSize`.
Other links (Linkedin, GitHub) *are* defined with `em`, so they *do*
behave correctly.

For the moment, I'm only fixing Bluesky because that's what I use; I
might get around to fixing the others if there is a need (or other users
find their way to my fork).