Code backgrounds should always be dark
Because the rest of the code highlight theme doesn't change with
dark/light modes *and* because those colors are only visible against a
dark background; the `background-color` for `pre` and `code` blocks
should *always* be dark.

Added a `hugo.yaml` configuration ooption `colors.codeBackgroundColor`
to manually change this color.
Added padding to code blocks in the `#recent-posts` and `#list-page`
sections, so that the dark background looks better in light mode.