Adjust font sizes/styles in page/footer lists
The relative sizes of headers in the blog post previews (in both the
list pages and the recent posts footer) didn't make much semantic sense,
so they've been adjusted to both better show the hierarchy and have a
nicer look.

The font size for the previews has been reduced a bit to differentiate
them from main page content.