Use larger previews in list pages
Instead of a 3-column card layout for the list pages (like the blog
index), we're going for a single larger column layout throughout.

On larger screens we'll use somewhat narrower columns (like a
`col-9-lg` and `col-11-md`), just to provide a bit of breathing room on
the sides and avoid overly long text lines.

This also exposed a bug in the original css for `card-img-top`: namely
that the `object-fit: cover` was not set as it was for the recent-posts
partial. This lead to distorted, stretched images in the now wider
preview cards. This has been addressed.